Add an option to customise armeria's WebClient when creating the backend. Use in tests to disable the write timeout which caused failures for SSE tests.
