+++ armeria-backend/fs2-ce2/src/test/scala/sttp/client3/armeria/fs2/ArmeriaFs2StreamingTest.scala
@@ -2,12 +2,18 @@ package sttp.client3.armeria.fs2
 
 import cats.effect.IO
 import sttp.capabilities.fs2.Fs2Streams
-import sttp.client3.SttpBackend
+import sttp.client3.{SttpBackend, SttpBackendOptions}
+import sttp.client3.armeria.ArmeriaWebClient
 import sttp.client3.impl.fs2.Fs2StreamingTest
+
+import java.time.Duration
 
 class ArmeriaFs2StreamingTest extends Fs2StreamingTest {
   override val backend: SttpBackend[IO, Fs2Streams[IO]] =
-    ArmeriaFs2Backend()
+    ArmeriaFs2Backend.usingClient(
+      // the default caused timeouts in SSE tests
+      ArmeriaWebClient.newClient(SttpBackendOptions.Default, _.writeTimeout(Duration.ofMillis(0)))
+    )
 
   override protected def supportsStreamingMultipartParts: Boolean = false
 }
