+++ armeria-backend/src/main/scala/sttp/client3/armeria/future/ArmeriaFutureBackend.scala
@@ -4,7 +4,7 @@ import com.linecorp.armeria.client.WebClient
 import com.linecorp.armeria.common.HttpData
 import com.linecorp.armeria.common.stream.StreamMessage
 import org.reactivestreams.Publisher
-import sttp.client3.armeria.AbstractArmeriaBackend.newClient
+import sttp.client3.armeria.ArmeriaWebClient.newClient
 import sttp.client3.armeria.{AbstractArmeriaBackend, BodyFromStreamMessage}
 import sttp.client3.internal.NoStreams
 import sttp.client3.{FollowRedirectsBackend, SttpBackend, SttpBackendOptions}
