+++ armeria-backend/zio1/src/test/scala/sttp/client3/armeria/zio/ArmeriaZioStreamingTest.scala
@@ -1,7 +1,8 @@
 package sttp.client3.armeria.zio
 
 import sttp.capabilities.zio.ZioStreams
-import sttp.client3.SttpBackend
+import sttp.client3.armeria.ArmeriaWebClient
+import sttp.client3.{SttpBackend, SttpBackendOptions}
 import sttp.client3.impl.zio.{ZioServerSentEvents, ZioTestBase}
 import sttp.client3.internal._
 import sttp.client3.testing.ConvertToFuture
@@ -10,11 +11,18 @@ import sttp.model.sse.ServerSentEvent
 import zio.stream.Stream
 import zio.{Chunk, Task}
 
+import java.time.Duration
+
 class ArmeriaZioStreamingTest extends StreamingTest[Task, ZioStreams] with ZioTestBase {
   override val streams: ZioStreams = ZioStreams
 
   override val backend: SttpBackend[Task, ZioStreams] =
-    runtime.unsafeRun(ArmeriaZioBackend())
+    runtime.unsafeRun(
+      ArmeriaZioBackend.usingClient(
+        // the default caused timeouts in SSE tests
+        ArmeriaWebClient.newClient(SttpBackendOptions.Default, _.writeTimeout(Duration.ofMillis(0)))
+      )
+    )
   override implicit val convertToFuture: ConvertToFuture[Task] = convertZioTaskToFuture
 
   override def bodyProducer(arrays: Iterable[Array[Byte]]): Stream[Throwable, Byte] =
